+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+- Dynamic Report: fields can now be used in multiple zones simultaneously (rows + filters, columns + filters)
+- Dynamic Report: right-click on a filter value to exclude it (shown with strikethrough in red)
+
 ## [0.3.9]
 
 ### Added